+++ src/renderer/src/providers/AppProviders.tsx
@@ -6,6 +6,8 @@ import { I18nextProvider } from 'react-i18next'
 import { HashRouter } from 'react-router-dom'
 import i18n from '../i18n'
 import App from '../App'
+import CheckNotificationSync from '../components/CheckNotificationSync'
+import CheckToastHost from '../components/CheckToastHost'
 import NativeTitleBarSync from '../components/NativeTitleBarSync'
 import ProxyDataSync from '../components/ProxyDataSync'
 import { useSettingsStore } from '../store/settingsStore'
@@ -49,6 +51,8 @@ function AppProviders(): React.JSX.Element {
           <ThemeModeSync />
           <NativeTitleBarSync />
           <ProxyDataSync />
+          <CheckNotificationSync />
+          <CheckToastHost />
           <HashRouter>
             <App />
           </HashRouter>